bhutanese_currency_detector/requirements.txt

```
Django==4.2.7
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
```